Date handling for mpoxseir

We need to map "dates" onto dust::dust's concept of model "step" and we do this by mapping a date such as 2024-03-02 into the number of days since the start of 2023 (426 here, with the 1st of January being day 0). We call this integer number an "mpoxseir date".

mpoxseir_date(date)

mpoxseir_date_as_date(date)

as_mpoxseir_date(date)

as_date(date)

Arguments

date

A Date object, or something that can be converted to one, or an "mpoxseir date"; see Details

Value

An integer, being the number of days into 2023

Details

There are several related functions here

mpoxseir_date converts its argument into an R Date object, then applies this tranformation. If the argument is not a Date object or a string representing one, an error will be thrown.

mpoxseir_date_to_date does the reverse conversion to mpoxseir_date, converting an integer mpoxseir date into an R Date

as_mpoxseir_date does the same conversion as mpoxseir_date but will assume that an integer already represents an mpoxseir date and will return it unmodified rather than erroring.

as_date does a string to date conversion, using as.Date() but requiring the dates are in ISO 8601 (YYYY-MM-DD) format (it is a helper that avoids conversion to NA, instead throwing an error)

Examples

# Convert dates into mpoxseir dates:
mpoxseir::mpoxseir_date("2023-01-01")
#> [1] 0
mpoxseir::mpoxseir_date(c("2024-03-01", "2024-10-01"))
#> [1] 425 639

# Reverse the conversion:
mpoxseir::mpoxseir_date_as_date(0)
#> [1] "2023-01-01"
mpoxseir::mpoxseir_date_as_date(c(425, 639))
#> [1] "2024-03-01" "2024-10-01"

# Double conversion not possible with mpoxseir_date...
try(mpoxseir::mpoxseir_date(61))
#> Error in as_date(date) : Expected ISO dates or R dates - please convert
# ...but allowed with as_mpoxseir_date
mpoxseir::as_mpoxseir_date(61)
#> [1] 61

# Strict date conversion with as_date
mpoxseir::as_date("2024-03-01")
#> [1] "2024-03-01"
try(mpoxseir::as_date("03-01-2024"))
#> Error in mpoxseir::as_date("03-01-2024") : 
#>   Expected ISO dates or R dates - please convert