Run the model with single set of parameter values

Run the model with single set of parameter values

run_mpoxSEIR_targetedVax_single(
  region,
  initial_infections,
  n_weeks,
  R0_hh,
  R0_sw_st,
  beta_z_max,
  n_vax = 0,
  daily_doses = 0,
  N_prioritisation_steps = 0,
  prioritisation_strategy = 0,
  vaccination_coverage_target = 0,
  vaccine_uptake = 0,
  ve_T = 0,
  ve_I = 0,
  ve_D = 0,
  vaccination_campaign_length = 0,
  overrides = list(),
  n_particles = 1,
  n_threads = 1,
  seed = 42,
  deterministic = TRUE,
  outputs_retained = NULL
)

Arguments

region

The region to run the model for, must be either "equateur" or "sudkivu"

initial_infections

The initial number of infections to seed with

n_weeks

number of weeks to run for

R0_hh

R0 for the household

R0_sw_st

R0 for sex workers to people who buy sex

beta_z_max

beta for the age-group with highest zoonotic transmission (a number)

n_vax

number of vaccination compartments (integer, basis for an additional dimension in odin states)

daily_doses

the daily number of doses administered (matrix of vaccination_campaign_length * number of vaccination compartments)

N_prioritisation_steps

the number of different vaccination prioritisation categories we're considering

prioritisation_strategy

what each step corresponds to in terms of strategy (matrix of n_group * N_prioritisation_steps, with 1s and 0s indicating whether a group is included in prioritisation step)

vaccination_coverage_target

vaccination coverage target for each group and prioritisation step (matrix of n_group * N_prioritisation_steps)

vaccine_uptake

max achievable coverage for each group (vector of length n_group)

ve_T

vaccine efficacy against onwards transmissibility for each vaccinated compartment (vector of length n_vax)

ve_I

vaccine efficacy against infection for each vaccinated compartment (vector of length n_vax)

ve_D

vaccine efficacy against death for each vaccinated compartment (vector of length n_vax)

vaccination_campaign_length

length of the vaccination campaign (in timesteps NOT days - CHECK WITH RUTH THIS IS RIGHT)

overrides

list of other model parameters which if specified will overwrite the defaults

n_particles

Number of particles

n_threads

Number of threads

seed

The random seed to use

deterministic

Logical, whether to run the model deterministically or not

outputs_retained

The outputs to retain

Value

The output from running the model